backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.3",
    "python-pptx>=1.0.2",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
test = [
    "coverage>=7.13.1",
    "pytest>=9.0.2",
]
